requirements.txt
Markdown>=3.3.7
PyYAML>=6.0
Pygments>=2.12.0
numpy>=1.23.3
scour>=0.38.2
requests>=2.28.2